# openapi.yaml
openapi: 3.0.3
info:
  title: k2pbot — Generic Chatbot Platform API
  description: |
    REST API for the k2pbot Generic Chatbot Platform built on Spring Boot / Spring AI.
    Provides blocking chat, Server-Sent Events streaming, session management, and admin
    configuration endpoints.

    **Error codes returned in `ApiErrorResponse.errorCode`**
    | Code | Meaning |
    |---|---|
    | `VALIDATION_ERROR` | Bean-validation failure on the request body |
    | `ASSISTANT_NOT_FOUND` | No active assistant found for the given code |
    | `CONFIG_INVALID` | Assistant config failed validation (422) |
    | `CONFIG_NOT_FOUND` | Config resolution / knowledge-base lookup failed |
    | `INVALID_REQUEST` | Unauthorised runtime override or bad argument |
    | `TOOL_NOT_ALLOWED` | Policy violation during tool execution (422) |
    | `STREAMING_NOT_SUPPORTED` | Assistant does not support streaming |
    | `MODEL_UNAVAILABLE` | Target LLM model is unavailable (503) |
    | `LLM_PROVIDER_ERROR` | Upstream LLM provider error (502) |
    | `INTERNAL_ERROR` | Unexpected server error or execution failure |
  version: "1.0.0"
  contact:
    email: kaushik117@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/admin/tenant-overrides:
    get:
      tags: [AdminTenantOverrides]
      operationId: listTenantOverrides
      summary: List tenant overrides
      description: Returns a paginated list of tenant-specific config overrides.
      parameters:
        - name: tenantId
          in: query
          required: false
          schema:
            type: string
        - name: assistantCode
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          description: Tenant override list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantOverrideDto"
    post:
      tags: [AdminTenantOverrides]
      operationId: createTenantOverride
      summary: Create tenant override
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantOverrideRequest"
      responses:
        "201":
          description: Tenant override created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantOverrideDto"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/admin/tenant-overrides/{id}:
    get:
      tags: [AdminTenantOverrides]
      operationId: getTenantOverride
      summary: Get tenant override
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: Tenant override found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantOverrideDto"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [AdminTenantOverrides]
      operationId: updateTenantOverride
      summary: Update tenant override
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantOverrideRequest"
      responses:
        "200":
          description: Tenant override updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantOverrideDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [AdminTenantOverrides]
      operationId: deleteTenantOverride
      summary: Delete tenant override
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "204":
          description: Tenant override deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Admin — RAG Audit ──────────────────────────────────────────────────────
...
components:
...
  parameters:
...
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Numeric primary key of the resource
    pageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Zero-based page number
    sizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 200
      description: Page size

  # ── Reusable Responses ─────────────────────────────────────────────────────
  responses:
    BadRequest:
      description: "400 — Validation failure or bad argument"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    NotFound:
      description: "404 — Resource not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
  schemas:
...
    TenantOverrideDto:
      type: object
      description: Tenant-specific config override record
      properties:
        id:
          type: integer
          format: int64
          example: 20
        tenantId:
          type: string
          example: rahi-prod
        assistantCode:
          type: string
          example: loan-advisor
        overrideType:
          type: string
          description: The config section being overridden (e.g. `memory`, `rag`, `response`).
          example: memory
        overridePayloadJson:
          type: string
          description: JSON-serialized override payload applied on top of the base config.
          example: '{"messageWindowSize":10}'
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2026-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-31T09:55:00Z"
        createdBy:
          type: string
          nullable: true
          example: admin
    CreateTenantOverrideRequest:
      type: object
      required:
        - tenantId
        - assistantCode
        - overrideType
        - overridePayloadJson
      properties:
        tenantId:
          type: string
          maxLength: 100
          example: rahi-prod
        assistantCode:
          type: string
          maxLength: 100
          example: loan-advisor
        overrideType:
          type: string
          description: The config section being overridden.
          example: memory
        overridePayloadJson:
          type: string
          description: JSON-serialized override payload.
          example: '{"messageWindowSize":10}'
    UpdateTenantOverrideRequest:
      type: object
      properties:
        overrideType:
          type: string
          nullable: true
          example: memory
        overridePayloadJson:
          type: string
          nullable: true
          example: '{"messageWindowSize":5}'
        active:
          type: boolean
          nullable: true

    # ── Admin — RAG Audit DTOs ────────────────────────────────────────────────
...
    ApiErrorResponse:
      type: object
      description: Standardised error envelope returned for all error responses
      properties:
        errorCode:
          type: string
          description: Machine-readable error code (see API description for full list).
          example: VALIDATION_ERROR
        errorMessage:
          type: string
          description: Human-readable description of the error.
          example: "Request validation failed"
        requestId:
          type: string
          description: Correlation ID from MDC — include in support tickets.
          example: req-7f3a9c12
        timestamp:
          type: string
          format: date-time
          example: "2026-05-31T10:00:00Z"
        fieldErrors:
          type: array
          description: Per-field validation errors (only present on 400 VALIDATION_ERROR).
          items:
            $ref: "#/components/schemas/FieldValidationError"
    FieldValidationError:
      type: object
      description: A single field-level validation failure
      properties:
        field:
          type: string
          description: The request body field path that failed validation.
          example: message
        message:
          type: string
          description: Validation constraint message.
          example: "must not be blank"

    # ── Enumerations ──────────────────────────────────────────────────────────
...